Fix Conditional Tag if not for a network admin page
git-svn-id: https://plugins.svn.wordpress.org/image-slider-widget/trunk@1069538 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ easy-slider-widget-lite.php
@@ -194,7 +194,7 @@ register_activation_hook( __FILE__, 'ewic_plugin_activate' );
 
 function ewic_load_plugin() {
 
-    if ( is_admin() && get_option( 'activatedewic' ) == 'ewic-activate' ) {
+    if ( is_admin() && get_option( 'activatedewic' ) == 'ewic-activate' && !is_network_admin() ) {
 		delete_option( 'activatedewic' );
 		wp_redirect("edit.php?post_type=easyimageslider&page=ewiccomparison");
